v0.16 переделки и фиксы...

+++ backend/api/views.py
@@ -1,36 +1,36 @@
-from django.http.response import HttpResponse
-from djoser.views import UserViewSet
 from django.contrib.auth import get_user_model
+from django.db.models import F, Sum
+from django.http.response import HttpResponse
 from django_filters.rest_framework import DjangoFilterBackend
+from djoser.views import UserViewSet
 from rest_framework import status
-from rest_framework.viewsets import ModelViewSet, ReadOnlyModelViewSet
-from rest_framework.generics import get_object_or_404
 from rest_framework.decorators import action
+from rest_framework.generics import get_object_or_404
 from rest_framework.permissions import AllowAny, IsAuthenticated
 from rest_framework.response import Response
+from rest_framework.viewsets import ModelViewSet, ReadOnlyModelViewSet
 
 from .filters import IngredientSearchFilter, RecipeFilter
 from .pagination import ApiPagination
 from .permissions import IsAuthorOrReadOnly
 from .serializers import (
     CustomUserSerializer,
+    IngredientSerializer,
     RecipeSerializer,
+    RepresentationRecipeSerializer,
     SubscribeSerializer,
     SubscriptionsSerializer,
-    FavoriteSerializer,
-    ShoppingCartSerializer,
-    IngredientSerializer,
     TagSerializer,
 )
-from recipes.models import (
-    Recipe,
-    Ingredient,
-    Tag,
-    Follow,
-    Favorite,
-    RecipeIngredientAmount,
-)
 from cart.models import ShoppingCart
+from recipes.models import (
+    Favorite,
+    Follow,
+    Ingredient,
+    Recipe,
+    RecipeIngredientAmount,
+    Tag,
+)
 
 User = get_user_model()
 
@@ -42,39 +42,33 @@ class CustomUserViewSet(UserViewSet):
     queryset = User.objects.all()
     pagination_class = ApiPagination
     serializer_class = CustomUserSerializer
-    lookup_field = "id"
 
     @action(
         detail=True,
-        methods=["post"],
+        methods=["post", "delete"],
         permission_classes=[IsAuthenticated],
     )
-    def subscribe(self, request, id=None):
+    def subscribe(self, request, *args, **kwargs):
         user = self.request.user
-        author = get_object_or_404(User, id=id)
-        data = {
-            "user": user.id,
-            "author": author.id,
-        }
-        serializer = SubscribeSerializer(
-            data=data, context={"request": request}
-        )
-        serializer.is_valid(raise_exception=True)
-        serializer.save()
-        queryset = get_object_or_404(Follow, user=user)
-        serializer = SubscriptionsSerializer(
-            queryset, context={"request": request}
-        )
-
-        return Response(serializer.data, status=status.HTTP_201_CREATED)
-
-    @subscribe.mapping.delete
-    def delete_subscribe(self, request, id=None):
-        user = self.request.user
-        author = get_object_or_404(User, id=id)
-        subscribe = get_object_or_404(Follow, user=user, author=author)
-        subscribe.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        author = self.get_object()
+        if request.method == "POST":
+            data = {
+                "user": user.id,
+                "author": author.id,
+            }
+            serializer = SubscribeSerializer(
+                data=data, context={"request": request}
+            )
+            serializer.is_valid(raise_exception=True)
+            serializer.save()
+            serializer = SubscriptionsSerializer(
+                author, context={"request": request}
+            )
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        else:
+            subscribe = get_object_or_404(Follow, user=user, author=author)
+            subscribe.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
 
     @action(
         detail=False,
@@ -83,7 +77,7 @@ class CustomUserViewSet(UserViewSet):
     )
     def subscriptions(self, request):
         user = self.request.user
-        queryset = Follow.objects.filter(user=user)
+        queryset = User.objects.filter(follow_author__user=user)
         paginated_follow = self.paginate_queryset(queryset=queryset)
         serializer = SubscriptionsSerializer(
             paginated_follow, many=True, context={"request": request}
@@ -104,64 +98,45 @@ class RecipeViewSet(ModelViewSet):
     def perform_create(self, serializer):
         return serializer.save(author=self.request.user)
 
+    def _do_post_delete(self, request, model):
+        user = self.request.user
+        recipe = self.get_object()
+        if request.method == "POST":
+            if model.objects.filter(
+                user=user,
+                recipe=recipe,
+            ).exists():
+                return Response(status=status.HTTP_400_BAD_REQUEST)
+            model.objects.create(
+                user=user,
+                recipe=recipe,
+            )
+            serializer = RepresentationRecipeSerializer(
+                recipe,
+                context={"request": request},
+            )
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+
+        else:
+            item = get_object_or_404(model, user=user, recipe=recipe)
+            item.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+
     @action(
-        methods=["post"],
+        methods=["post", "delete"],
         permission_classes=[IsAuthenticated],
         detail=True,
     )
     def favorite(self, request, pk=None):
-        user = self.request.user
-        recipe = get_object_or_404(Recipe, id=pk)
-        data = {
-            "user": user.id,
-            "favorite": recipe.id,
-        }
-        serializer = FavoriteSerializer(
-            data=data, context={"request": request}
-        )
-        serializer.is_valid(raise_exception=True)
-        serializer.save()
-        return Response(serializer.data, status=status.HTTP_201_CREATED)
-
-    @favorite.mapping.delete
-    def delete_favorite(self, request, pk=None):
-        user = self.request.user
-        recipe = get_object_or_404(Recipe, id=pk)
-        favorite = get_object_or_404(Favorite, user=user, favorite=recipe)
-        favorite.delete()
-
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        return self._do_post_delete(request, Favorite)
 
     @action(
-        methods=["post"],
-        detail=True,
+        methods=["post", "delete"],
         permission_classes=[IsAuthenticated],
+        detail=True,
     )
     def shopping_cart(self, request, pk=None):
-        user = self.request.user
-        recipe = get_object_or_404(Recipe, id=pk)
-
-        data = {
-            "user": user.id,
-            "recipe": recipe.id,
-        }
-        serializer = ShoppingCartSerializer(
-            data=data,
-            context={"request": request},
-        )
-        serializer.is_valid(raise_exception=True)
-        serializer.save()
-
-        return Response(serializer.data, status=status.HTTP_201_CREATED)
-
-    @shopping_cart.mapping.delete
-    def delete_shopping_cart(self, request, pk=None):
-        user = request.user
-        recipe = get_object_or_404(Recipe, id=pk)
-        favorites = get_object_or_404(ShoppingCart, user=user, recipe=recipe)
-        favorites.delete()
-
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        return self._do_post_delete(request, ShoppingCart)
 
     @action(
         methods=["get"],
@@ -170,31 +145,22 @@ class RecipeViewSet(ModelViewSet):
     )
     def download_shopping_cart(self, request):
         user = self.request.user
-        shopping_cart = user.cart_user.all()
         filename = f"{user.username}_shopping_cart.txt"
-        ingredients_list = {}
-        for recipes_list in shopping_cart:
-            recipe = recipes_list.recipe
-            ingredients = RecipeIngredientAmount.objects.filter(recipe=recipe)
-            for i in ingredients:
-                amount = i.amount
-                name = i.ingredient.name
-                measurement_unit = i.ingredient.measurement_unit
-                if name not in ingredients_list:
-                    ingredients_list[name] = {
-                        "measurement_unit": measurement_unit,
-                        "amount": amount,
-                    }
-                else:
-                    ingredients_list[name]["amount"] = (
-                        ingredients_list[name]["amount"] + amount
-                    )
 
-        shopping_list = []
-        for i in ingredients_list:
-            shopping_list.append(
-                f'{i} - {ingredients_list[i]["amount"]} '
-                f'{ingredients_list[i]["measurement_unit"]} \n'
+        ingredients = (
+            RecipeIngredientAmount.objects.filter(
+                recipes__shopping_cart__user=user
+            )
+            .values(
+                ingr_name=F("ingredient__name"),
+                unit=F("ingredients__measurement_unit"),
+            )
+            .annotate(amount_sum=Sum("amount"))
+        )
+        shopping_list = ()
+        for ingr in ingredients:
+            shopping_list += (
+                f'{ingr["ingr_name"]}: {ingr["amount_sum"]} {ingr["unit"]}\n'
             )
 
         response = HttpResponse(
